% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{calculateEntropyMatrix}
\alias{calculateEntropyMatrix}
\title{Calculate entropy matrix}
\usage{
calculateEntropyMatrix(matrixWeighted)
}
\description{
This function calculates a matrix of entropy values given a matrix.
}
