% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{calculateEntropyMatrixAllEdge}
\alias{calculateEntropyMatrixAllEdge}
\title{Calculate entropy matrix for all edges}
\usage{
calculateEntropyMatrixAllEdge(matrixWeighted)
}
\arguments{
\item{matrixWeighted}{input matrix}
}
\description{
This function calculates entropy values for all edges in the given matrix.
}
