% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_distance_matrix.R
\name{createCNVMatrix}
\alias{createCNVMatrix}
\title{Create CNV Distance Matrix}
\usage{
createCNVMatrix(
  rawCountsMatrix,
  annotations,
  geneOrder,
  ref_sample,
  outputFile
)
}
\arguments{
\item{rawCountsMatrix}{Name of raw counts matrix file}

\item{annotations}{Name of annotations file}

\item{geneOrder}{Name of gene order file}

\item{outputFile}{Name of output file that inferCNV object is saved to}
}
\description{
This function creates a cell-cell distance matrix using CNV data.
}
