% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_heatmap.R
\name{createComplexHeatmap}
\alias{createComplexHeatmap}
\title{Create SNV Distance Matrix input from VCF File}
\usage{
createComplexHeatmap(
  snvDistMatrix,
  cnvDistMatrix,
  integratedDistMatrix,
  infercnvObj,
  geneOrder
)
}
\arguments{
\item{snvDistMatrix}{cell-cell distance matrix generated with SNV data}

\item{cnvDistMatrix}{cell-cell distance matrix generated with CNV data}

\item{integratedDistMatrix}{integrated distance matrix with CNV and SNV data}

\item{infercnvObj}{RDA File with inferCNV object}

\item{geneOrder}{Name of geneOrder file, file with gene annotations}
}
\value{
interactive complex heatmap
}
\description{
This function finds chromosome information, alternative allele read counts (X),
total read counts (N), and mutation profile matrix (Z) from VCF file of mutations
}
