% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{createIntegratedMatrix}
\alias{createIntegratedMatrix}
\title{Create Integrated Matrix}
\usage{
createIntegratedMatrix(cnvDistMatrix, snvDistMatrix)
}
\arguments{
\item{cnvDistMatrix}{cell-cell distance matrix generated with CNV data}

\item{snvDistMatrix}{cell-cell distance matrix generated with SNV data}
}
\value{
distcombined, integrated matrix
}
\description{
This function creates the integrated CNV and SNV cell-cell distance matrix
using an entropy weighted methods.
}
