% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_distance_matrix.R
\name{createSNVMatrix}
\alias{createSNVMatrix}
\title{Create Distance Matrix of SNVs}
\usage{
createSNVMatrix(filevcf, mutations, SRR_ID, output_csv)
}
\arguments{
\item{filevcf}{Path to input VCF file}

\item{mutations}{Path to input mutations RDA file}

\item{SRR_ID}{Path to file with SRR IDs}

\item{output_csv}{Name of output csv file with distance matrix}
}
\value{
SNV Distance Matrix
}
\description{
This function creates a distance matrix of SNVs using alternative allele
read counts, total allele read counts, and a mutation profile matrix
}
