% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{EntropytoALPHA}
\alias{EntropytoALPHA}
\title{Calculate alpha values}
\usage{
EntropytoALPHA(entropylist, theta)
}
\arguments{
\item{entropylist}{list of entropy values}

\item{theta}{number of networks to be integrated}
}
\description{
This function calculates alpha values given theta and a list of entropy values.
}
