% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_distance_matrix.R
\name{makeInferCNVObject}
\alias{makeInferCNVObject}
\title{Create InferCNV Object}
\usage{
makeInferCNVObject(rawCountsMatrix, annotations, geneOrder, outputFile)
}
\arguments{
\item{rawCountsMatrix}{Name of raw counts matrix file}

\item{annotations}{Name of annotations file}

\item{geneOrder}{Name of gene order file}

\item{outputFile}{Name of output file that inferCNV object is saved to}
}
\description{
This function creates an inferCNV object.
}
