% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_distance_matrix.R
\name{vcf_to_SNVinput}
\alias{vcf_to_SNVinput}
\title{Create SNV Distance Matrix input from VCF File}
\usage{
vcf_to_SNVinput(filevcf)
}
\arguments{
\item{filevcf}{Path to input VCF file}
}
\value{
List with Mutation Information (Info, X, N, Z)
}
\description{
This function finds chromosome information, alternative allele read counts (X),
total read counts (N), and mutation profile matrix (Z) from VCF file of mutations
}
